+++ vite.config.js
@@ -13,10 +13,10 @@ const config = {
     VITE_APP_VERSION: JSON.stringify(version),
     VITE_IC_FEED_CANISTER_SECRET: JSON.stringify(process.env.FEED_SECRET)
   },
-  build: { target: "es2020" },
+  build: {target: 'es2020'},
   optimizeDeps: {
-    esbuildOptions: { target: "es2020", supported: { bigint: true } },
-  },
+    esbuildOptions: {target: 'es2020', supported: {bigint: true}}
+  }
 };
 
 export default config;